requirements.txt
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
Flask==0.10.1
Flask-Cors==3.0.3
Flask-GraphQL==1.3.0
graphene==2.1.2
graphene-sqlalchemy==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
idna==2.6
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
promise==2.1
psycopg2==2.7.3.2
requests==2.18.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.5
typing==3.6.4
urllib3==1.22
Werkzeug==0.14.1
